# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FPA Results API",
    "version": "0.1.0",
    "description": "Read-only API over the freestyle frisbee judging, results and rankings services.\n\nUpstream is five independent microservices, each with its own store and no shared query\nlayer. This API fetches all of them, resolves duplicate player identities, normalizes\ninconsistent division names, and joins ranking points back to the events that produced\nthem — then serves the result from an in-memory index.\n\n**Freshness.** Data is refreshed on a tiered schedule rather than per request, so responses\ncan be up to an hour behind upstream. `GET /health` reports exactly how stale the data is.\n\n**Data quality.** Upstream enforces no referential integrity. Records that reference a\nmissing player or event are kept and flagged (`unknown: true` on a team member, warnings\non `/health`) rather than dropped, so nothing disappears silently."
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Shared secret from REFRESH_TOKEN. Only used by POST /refresh."
      }
    },
    "schemas": {
      "PaginatedEvents": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventSummary"
            }
          },
          "total": {
            "type": "number",
            "description": "Total matching items, before pagination"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "EventSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": [
              "string",
              "null"
            ],
            "description": "YYYY-MM-DD"
          },
          "endDate": {
            "type": [
              "string",
              "null"
            ],
            "description": "YYYY-MM-DD"
          },
          "createdAt": {
            "type": [
              "number",
              "null"
            ]
          },
          "fpaWebsiteId": {
            "type": [
              "string",
              "null"
            ]
          },
          "fpaWebsiteSlug": {
            "type": [
              "string",
              "null"
            ]
          },
          "divisions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Normalized divisions that have results"
          },
          "resultCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "startDate",
          "endDate",
          "createdAt",
          "fpaWebsiteId",
          "fpaWebsiteSlug",
          "divisions",
          "resultCount"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "number"
          }
        },
        "required": [
          "error",
          "status"
        ]
      },
...
      "PaginatedPlayers": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Player"
            }
          },
          "total": {
            "type": "number",
            "description": "Total matching items, before pagination"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "Player": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Canonical player GUID, shared across all upstream services"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "country": {
            "type": [
              "string",
              "null"
            ],
            "description": "3-letter country code"
          },
          "gender": {
            "type": [
              "string",
              "null"
            ],
            "description": "Withheld unless EXPOSE_PERSONAL_FIELDS is enabled"
          },
          "membership": {
            "type": [
              "number",
              "null"
            ],
            "description": "Withheld unless EXPOSE_PERSONAL_FIELDS is enabled"
          },
          "fpaWebsiteId": {
            "type": [
              "string",
              "null"
            ],
            "description": "Id of this player on the legacy FPA website"
          },
          "aliasIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Duplicate GUIDs that resolve to this player"
          },
          "createdAt": {
            "type": [
              "number",
              "null"
            ]
          },
          "lastActive": {
            "type": [
              "number",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "fullName",
          "country",
          "gender",
          "membership",
          "fpaWebsiteId",
          "aliasIds",
          "createdAt",
          "lastActive"
        ]
      },
...
      "PaginatedRankings": {
        "type": "object",
        "properties": {
          "series": {
            "type": "string"
          },
          "division": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RankingEntry"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "series",
          "division",
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
...
      "PaginatedRatings": {
        "type": "object",
        "properties": {
          "series": {
            "type": "string"
          },
          "division": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RatingEntry"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "series",
          "division",
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
...
      "SnapshotList": {
        "type": "object",
        "properties": {
          "snapshots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SnapshotRef"
            }
          }
        },
        "required": [
          "snapshots"
        ]
      },
...
      "EventMatches": {
        "type": "object",
        "properties": {
          "query": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "startDate": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "name",
              "startDate"
            ]
          },
          "candidates": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "event": {
                  "$ref": "#/components/schemas/EventSummary"
                },
                "score": {
                  "type": "number",
                  "description": "Combined score, 0 to 1"
                },
                "nameScore": {
                  "type": "number"
                },
                "dateScore": {
                  "type": "number"
                },
                "confident": {
                  "type": "boolean",
                  "description": "Near-certain match, but still requires confirmation"
                }
              },
              "required": [
                "event",
                "score",
                "nameScore",
                "dateScore",
                "confident"
              ]
            }
          }
        },
        "required": [
          "query",
          "candidates"
        ]
      },
      "PlayerMatches": {
        "type": "object",
        "properties": {
          "query": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              }
            },
            "required": [
              "name"
            ]
          },
          "candidates": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "player": {
                  "$ref": "#/components/schemas/Player"
                },
                "score": {
                  "type": "number"
                },
                "placementCount": {
                  "type": "number",
                  "description": "Use as a tiebreaker: an active player is the likelier match"
                },
                "confident": {
                  "type": "boolean"
                }
              },
              "required": [
                "player",
                "score",
                "placementCount",
                "confident"
              ]
            }
          }
        },
        "required": [
          "query",
          "candidates"
        ]
      },
...
    },
...
  },
  "paths": {
    "/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "List events",
        "description": "Every event known to the upstream event-summary service, newest first. Events without a start date sort last rather than being omitted.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 50,
              "description": "Maximum items to return",
              "example": 50
            },
            "required": false,
            "description": "Maximum items to return",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "default": 0,
              "description": "Items to skip",
              "example": 0
            },
            "required": false,
            "description": "Items to skip",
            "name": "offset",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Case-insensitive event name search"
            },
            "required": false,
            "description": "Case-insensitive event name search",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2026-07-25",
              "description": "Only events starting on or after this date"
            },
            "required": false,
            "description": "Only events starting on or after this date",
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2026-07-25",
              "description": "Only events starting on or before this date"
            },
            "required": false,
            "description": "Only events starting on or before this date",
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Only events with results in this division",
              "example": "Open Pairs"
            },
            "required": false,
            "description": "Only events with results in this division",
            "name": "division",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Restrict to events that have results"
            },
            "required": false,
            "description": "Restrict to events that have results",
            "name": "hasResults",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEvents"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Data is still loading or upstream is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/players": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "List players",
        "description": "Only canonical players are listed. Upstream stores duplicate entries for the same person linked by `aliasKey`; those are merged here and their GUIDs appear in `aliasIds`.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 50,
              "description": "Maximum items to return",
              "example": 50
            },
            "required": false,
            "description": "Maximum items to return",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "default": 0,
              "description": "Items to skip",
              "example": 0
            },
            "required": false,
            "description": "Items to skip",
            "name": "offset",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Case-insensitive player name search"
            },
            "required": false,
            "description": "Case-insensitive player name search",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 3,
              "description": "3-letter country code",
              "example": "ITA"
            },
            "required": false,
            "description": "3-letter country code",
            "name": "country",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Restrict to players with at least one recorded placement"
            },
            "required": false,
            "description": "Restrict to players with at least one recorded placement",
            "name": "hasResults",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of players",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPlayers"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Data is still loading or upstream is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/rankings": {
      "get": {
        "tags": [
          "Rankings"
        ],
        "summary": "Get the current ranking standings",
        "description": "Cumulative tournament points. Defaults to `ranking-open`. Each entry carries a `breakdown` tying its points back to the specific events that earned them — a join the upstream services do not offer.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100,
              "example": 100
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "default": 0,
              "description": "Items to skip",
              "example": 0
            },
            "required": false,
            "description": "Items to skip",
            "name": "offset",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Series key, e.g. \"ranking-open\". Overrides `division`.",
              "example": "ranking-open"
            },
            "required": false,
            "description": "Series key, e.g. \"ranking-open\". Overrides `division`.",
            "name": "series",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Short division token, e.g. \"open\" or \"women\". Defaults to \"open\".",
              "example": "open"
            },
            "required": false,
            "description": "Short division token, e.g. \"open\" or \"women\". Defaults to \"open\".",
            "name": "division",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of ranked players",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRankings"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Data is still loading or upstream is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/ratings": {
      "get": {
        "tags": [
          "Ratings"
        ],
        "summary": "Get the current rating standings",
        "description": "Ratings are an Elo-style strength estimate derived from head-to-head outcomes. They are NOT the same as rankings: rankings accumulate tournament points over a season, ratings estimate current strength and have no points breakdown. A rating backed by a low `matchCount` is statistically weak — surface that number alongside the rating rather than the rating alone. Defaults to `rating-open`.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100,
              "example": 100
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "default": 0,
              "description": "Items to skip",
              "example": 0
            },
            "required": false,
            "description": "Items to skip",
            "name": "offset",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Series key, e.g. \"rating-open\". Overrides `division`.",
              "example": "rating-open"
            },
            "required": false,
            "description": "Series key, e.g. \"rating-open\". Overrides `division`.",
            "name": "series",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Short division token. Defaults to \"open\".",
              "example": "open"
            },
            "required": false,
            "description": "Short division token. Defaults to \"open\".",
            "name": "division",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "description": "Exclude players with fewer than this many head-to-head matches. Most rated players have very few, so a floor of 50-100 gives a far more meaningful leaderboard.",
              "example": 50
            },
            "required": false,
            "description": "Exclude players with fewer than this many head-to-head matches. Most rated players have very few, so a floor of 50-100 gives a far more meaningful leaderboard.",
            "name": "minMatchCount",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of rated players, strongest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRatings"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Data is still loading or upstream is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/snapshots": {
      "get": {
        "tags": [
          "Snapshots"
        ],
        "summary": "List historical ranking and rating snapshots",
        "description": "Rankings and ratings are published upstream as dated snapshots. This lists them newest first; snapshots flagged hidden upstream are excluded. Use these to chart a player or division over time.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "ranking",
                "rating"
              ]
            },
            "required": false,
            "name": "type",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "open"
            },
            "required": false,
            "name": "division",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Available snapshots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Data is still loading or upstream is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/match/events": {
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "Find upstream events that may correspond to one of ours",
        "description": "Results are suggestions ranked by similarity, never an automatic link. fpa-events and the upstream services share no identifier, so a confident-looking match can still be the wrong year or a junior division. Always require human confirmation before storing the id.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Event name to match against",
              "example": "German Championship 2026"
            },
            "required": true,
            "description": "Event name to match against",
            "name": "name",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2026-07-25",
              "description": "Improves ranking accuracy when supplied"
            },
            "required": false,
            "description": "Improves ranking accuracy when supplied",
            "name": "startDate",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "Restrict to events with results. Defaults to true."
            },
            "required": false,
            "description": "Restrict to events with results. Defaults to true.",
            "name": "hasResults",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked candidate events, best first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventMatches"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Data is still loading or upstream is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/match/players": {
      "get": {
        "tags": [
          "Matching"
        ],
        "summary": "Find upstream players that may correspond to one of our users",
        "description": "Results are suggestions ranked by similarity, never an automatic link. fpa-events and the upstream services share no identifier, so a confident-looking match can still be the wrong year or a junior division. Always require human confirmation before storing the id. Names are especially unreliable here — several people in the directory share a name.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Player name to match against",
              "example": "Francesco Santolin"
            },
            "required": true,
            "description": "Player name to match against",
            "name": "name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked candidate players, best first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerMatches"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Data is still loading or upstream is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}